+++ fits_storage/cal/calibration.py
@@ -383,6 +383,24 @@ class CalQuery(object):
         else:
             return self.raw().observation_type(name)
 
+    def raw_or_processed_by_types(self, name, processed):
+        """
+        Filter: If processed is ``True``, it is a shorthand for
+
+        ``reduction("PROCESSED_" + name)``
+
+        If not processed, then we look at the types (from AstroData tags).
+        Not all calibrations are noted in the OBSTYPE header and this is
+        an alternate way to distinguish them.
+
+        ``raw().filter(Header.types.like('%{0}%'.format(name)))``
+
+        """
+        if processed:
+            return self.reduction('PROCESSED_' + name)
+        else:
+            return self.raw().filter(Header.types.like('%''{0}''%'.format(name)))
+
     def bias(self, processed=False):
         """
         Filter: shorthand for ``raw_or_processed('BIAS', processed)``
@@ -482,6 +500,13 @@ class CalQuery(object):
             for key in kw:
                 ret = getattr(self, key)()
             return ret
+
+    def slitillum(self, processed=False):
+        """
+        Filter: shorthand for ``raw_or_processed('SLITILLUM', processed)``
+        """
+        return self.raw_or_processed_by_types('SLITILLUM', processed)
+
 
 class Calibration(object):
     """
@@ -656,3 +681,6 @@ class Calibration(object):
         # Not defined for this instrument
         return []
 
+    def slitillum(self, processed=False, howmany=None):
+        # Not defined for this instrument
+        return []